Using again my personal token, because the website-bot has been recognized as a non-human by github and the limit is only 60 requests per hour for him (the github plugin uses 1 request per file)
Here is the message from GitHub on hawkular-website profile:

One of our mostly harmless robots seems to think you are not a human.
Because of that, it’s hidden your profile from the public. If you really
are human, please contact support to have your profile reinstated. We
promise we won’t require DNA proof of your humanity.

:-)